requirements.txt
fastapi
uvicorn
httpx
python-dotenv
groq
google-generativeai
mistralai
tavily-python
elevenlabs
python-multipart